react_backend/requirement.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
pillow
numpy
django
django-environ
celery
redis
django-redis